refactor(server): Move PatternUtil to zally-core #560
