+++ server/src/main/java/de/zalando/zally/rule/zalando/NestedPathsMayBeRootPathsRule.kt
@@ -5,7 +5,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.PatternUtil.isPathVariable
+import de.zalando.zally.core.util.PatternUtil.isPathVariable
 
 @Rule(
     ruleSet = ZalandoRuleSet::class,
