+++ server/src/main/java/de/zalando/zally/rule/zalando/PluralizeResourceNamesRule.kt
@@ -10,7 +10,7 @@ import de.zalando.zally.rule.api.Context
 import de.zalando.zally.rule.api.Rule
 import de.zalando.zally.rule.api.Severity
 import de.zalando.zally.rule.api.Violation
-import de.zalando.zally.util.PatternUtil
+import de.zalando.zally.core.util.PatternUtil
 import de.zalando.zally.util.WordUtil.isPlural
 
 @Rule(
